refactor: cardInput 상태 관리 useCardInput 훅으로 대체
